Implement/apply "dx2(result,derivs)" function
This is related to this issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

This commit fully implements my new "dx2" function in all the AST node
classes of the expression library.  The dx2 function can serve as a
replacement for the original "val()" and "dx(int i)" functions.  My
commit from yesterday implemented the first version (which just combined
all the derivatives into a single traversal) and this one adds the
computation of the result to that traversal.

The premise for the new function is to perform all the necessary
calculations in a single AST traversal.  With the previous functions,
each quantity required a different traversal.  So, if you wanted to
evaluated an expression, and 3 of its derivatives, you would do 4
traversals; one for the result, and one for each derivative.

That wasn't the best way of doing things, but early in my development it
felt easier to debug.  Also, as most uses of the expression library
DON'T need derivatives, the consequences seemed low.

However, it did make circuits like the one attached to this bug, which
make very heavy use of behavioral models, much slower.  So, fixing it
was the right thing to do.  With this change, I have improved the
runtime of this circuit again.  Here is the latest report (again for the
first 5.0e-03 seconds of a 1sec simulation):

Xyce 7.1:  119.961 seconds    <- old expression library
Xyce 7.2:  227.651 seconds    <- new expression library, with limit op bug
Xyce 7.3:  235.599 seconds    <- new expression lib, limit op bug fixed
Xyce dev (5/21/2021):  230.936 seconds  <- group refactor for global_param
Xyce dev (5/24/2021):  179.048 seconds  <- Monday's changes
Xyce dev (5/25/2021):  157.217 seconds  <- Tuesday's first changes
Xyce dev (5/25/2021):  150.74 seconds <- Tuesday's second changes (first dx2)
Xyce dev (5/26/2021):  142.801 seconds <- Today's changes (second dx2)

So, implementing dx2 reduced the runtime from 157 to 142 seconds.
Combining all the derivatives (yesterday) took it from 157 seconds
to 150 seconds, and additionally combining the result (today) took
it from 150 seconds to 142.