Change the HB loader to handle mixed time and frequency domain sources
The sources were all loaded in time domain for HB analyses. This does
not work with the new frequency mapping based HB algorithm. The mapped
frequency and the time domain information are not related by FFT. The
AFM method is used to map the frequency so FFT can be applied in multi-tone
HB. But the method is used to evaluate the memoryless nonlinearity. The
sources in HB depend on time and the previous method to handle the source
evaluation does not work with the AFM algorithm.

The framework of loading sources in frequency domain is added recently. This
commit changes the HB loader to enable frequency domain loading of B vector and
time domain loading of F vecotr and Jacobian matrix for sources. This fits well with
the nonlinear/linear loading framework in HB. The B vectors can be loaded either
in time domain or frequency domain. But they should not be loaded in both.
With this change, the AFM algorithm now works for both single tone and multi-tone
HB circuits.

Xyce#160 #219