Refactor of the "outputs" group
This is related to this issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

This is also related more generally to my attempts to refactor and
improve the efficiency of the expression library.

I previously refactored the group classes so that they would be much
easier to optimize.  My goal at the time was primarily to fix the
"deviceExpressionGroup" class so that it would stop calling "find"
functions.   That update significantly improved the efficiency of Bsrc
devices, which rely on that deviceExpressionGroup.

With this commit, I'm now doing a similar refactor to the outputs group.
I expect that this will have a much less dramatic effect on performance,
as it only affects .PRINT line expressions.  But it is still the right
thing to do.

Prior to this commit, the outputs group was having to
allocate/deallocate various output "Ops" objects on the fly.  This
happened every time the group was invoked to populate values into an
expression.  These objects are small, and .PRINT line expressions
weren't that expensive to begin with (comapred to overall runtime), so
the cost of doing this wasn't that high.  But it is better to just
allocate the ops one time, and then reuse them throughout the
calculation.

In this commit I have done that.  I've created and populated
the functions "setupGroup" and "putValues", which were introduced by my
previous refactoring, but had not yet been implemented for this group.
With this commit all the expression variables that relied on "Ops" now
have their ops allocated once, rather than over and over again.