Very minor optimizations in the netlist parser
This commit is related to various performance issues.

This is related to this issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

Also:
Issue #35
"Reduce netlist setup time for modern PDKs"

The changes in this commit are small, but they shave a second off the
runtime of the BUG_1222_SON test case (the VCO circuit) from roughly 19
seconds to 17.x seconds.

I previously made the observation that during netlist parsing, the
expression library is doing more traversals than should be necessary.
This was often happening in the service of error traps, but that was a
very inefficient way to do it.    So, for example, parser would need to
know if an expression depended on lead currents, and to do this, it
would request a copy of the std::vector of lead currents, and then check
if this object was empty or not.  Often, assembling this vector required
a tree traversal, and this was expensive.  Also it was wasted effort,
because this vector would not then be used for anything.

To deal with this issue, I previously replaced most of these calls with
calls to booleans maintained by th expression library, and this was
faster.    However, I missed on of these calls, and it was showing up
when I ran the profiler on the BUG_1222_SON circuit.  Anyway, that is
fixed now.

The other thing changed in this commit was a
really tiny, probably inconsquential issue.  I noticed
that the parser would check expression strings to see if they were
surrounded by curly braces or not, and if they were, to strip them off.
This wasn't necessary with the new expression parser, since it already
understands curly braces in its grammar. So I removed this code in 3
places.  This may not change the performance very much, but it was a
useless bit of code at this point, so removing it was the
right thing to do.

NOTE: the tiny, inconsquential issue change (for curly braces) does
cause 7 "error exit" regression tests to fail.  These are not meaningful
failures.  They only happen because in some cases the output error
message now includes curly braces around the expression string, where
before they did not.