Remove redundant expression evals from the Bsrc
Issue #111
"Reduce redundant calls to "newExpression::evaluateFunction"
when using solution dependent expressions in Bsrcs"

Also, this is related to this issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

For various reasons, the Bsrc was evaluating expressions twice as much
as it needed to.  The reason for this was as follows:

1. Bsrc's use expressions.  The device is defined by either
I={expression} or V={expression}

2. The infrastructure of the device package considers the "I"
or "V" parameter of a Bsrc to be a "dependent" parameter.

3. This means, as a dependent parameter, that it will get updated
via that same infrastructure, in the device entity package, via a
call to newExpression::evaluateFunction.  This call evaluates
just the value of the expression, and not the derivatives.

4. But later, in the innards of the Bsrc, the function
newExpression::evaluate will be called for this same "I" or
"V" parameter.  The "evaluate" function produces both the
value of the expression and its derivatives.

5. So, this is redundant.  It should not be necessary to call
both "evaluateFunction" when "evaluate" is going to be called
later on the same expression.

It is possible that I introduced this redundancy when working on the new
expression library, since I refactored a lot of the "changed" logic.
Had to say at this point, partly because the "changed" logic was so hard
to follow.  But either way, the redundancy is now gone.

Note, there is one exception to this.  If the Bsrc expression contains
"ddt", then it actually does need to evaluate more than once.  So if
there are any ddt operators present in the expression, it is not put
into the exclude map.

So now, the development branch of Xyce is about as fast as Xyce 7.1 was,
for the customer circuit attached to the original issue, above.  As
before, these timings are for the first 5.0e-3 seconds of that circuit
(which originally specified a 1 sec transient).

Xyce 7.1:  119.961 seconds    <- old expression library
Xyce 7.2:  227.651 seconds    <- new expression library, with limit op bug
Xyce 7.3:  235.599 seconds    <- new expression lib, limit op bug fixed
Xyce dev (5/21/2021):  230.936 seconds  <- group refactor for global_param
Xyce dev (5/24/2021):  179.048 seconds  <- Monday's changes
Xyce dev (5/25/2021):  157.217 seconds  <- Tuesday's first changes
Xyce dev (5/25/2021):  150.74 seconds <- Tuesday's second changes (first dx2)
Xyce dev (5/26/2021):  142.801 seconds <- Wednesday's changes (second dx2)
Xyce dev (5/27/2021):  124.558 seconds <- Today's changes (SON issue 111)