Add dx2 support for most "function" operators
In my previous "dx2" work, I neglected to properly use dx2 in
this file.   This commit fixes that issue.    This should make
derivative evaluations for these functions (sin, cos, tan, etc) a little
faster.

This is related to this issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"