Add extra capability to "pre-evaluate" binary ops
This is related to this issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

It occured to me recently that any time the AST has a binary operation
involving two pure numbers, like 5.0*3.0, there is no reason to fully
allocate the binary operator, etc.  Instead, it is better to evaluate
the operation during parsing, and replace it with a number.  That way,
the same identical binary operation (multiplication, etc) won't be
wastefully performed over and over and over again.

This same sort of idea can easily be applied to many other operations,
but I haven't done that yet.  That will happen in future commits.