Make .STEP Work with .FOUR Lines
This commit enables .STEP to work with .FOUR lines.  A side effect is
that the default output file name for the non-step case has changed
from <netlistName>.four to <netlistName>.four0.  For the step case,
the output file names are <netlistName>.fourX where X is the step
number.  This naming convention is consistent with the Xyce output
file names for .MEASURE and .FFT lines.

Issue #228
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/228