Add more derivative tests
I noticed recently that the unit tests for derivatives didn't include
any division operators.  This corrects that, with one new test in both
the "double" and "complex" unit test programs.

Also, I noticed that a handful of derivative tests from the "double"
program had not been migrated over to the "complex" program yet.  This
commit does that as well.