reduce copies of RCPs in the parser
Issue #35
"Reduce netlist setup time for modern PDKs"

Issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

I noticed recently while running profiler tools that copying
reference-counted pointers (RCP) has an actual cost, performance-wise.
It isn't huge, but it is real.

In the expression parser, I was doing a lot of RCP copies needlessly.
There were various reasons that I did this, but part of it was that I
thought that, especially with good copiler optimizations, creating
temporary copies of RCP wouldn't cost anything (kind of like a temporary
copy of a conventional double* pointer gets optimized away).  But, you
probably can't optimize an RCP in the same way, as it has to do some
book keeping, so this thinking was wrong. This commit is an attempt
to fix that issue by removing some of the copies in the parser.  This
commit doesn't get rid of all of them, but it gets rid of maybe 90%.

This alone improved the parsing of the BUG_1222_SON test case (the VCO
example) by another second or two.  Back when I started my various
optimizations of the expression library, the "netlist import" phase for
that circuit took nearly 20 seconds on my laptop.  Now it takes more
like 11 seconds.  Before this set of changes it was around 13.