Add more pre-evaluation to various ops
Issue #35
"Reduce netlist setup time for modern PDKs"

Issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

This commit adds more code to "pre-evaluate" various operators when the
input is a pure number.  That way they are only computed once and are
simply represented by a number in the AST, rather than the original
operation.  So, for example, sin(number) gets replaced by that value.

This commit also includes more eliminations of RCP copies.