Optimize binary and function ops for const inputs
Issue #35
"Reduce netlist setup time for modern PDKs"

Issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

In light of my recent pre-compute changes, the impact of this is likely
to be small.  But, this commit updates the binary and function operators
so that if they are obviously linear w.r.t. an input, they won't waste
time computing the derivative.