Improve frequency domain output to work with AFM method
The real frequencies in the AFM based HB method is not sorted. The
mapped frequencies is sorted in order to work properly. The frequency
domain results which are in the same order  as the real frequencies are
not sorted. The results are correct, but the output of the positive and negative
frequency results are mixed. This commit changes the processing of
frequency domain output results. Both real and imaginary parts of
the results are handled properly. The frequency domain output has
similar ordering as the previous APFT based method.

Related to Xyce#160