Fix the initial guess calculation for AFM based HB method
The intial guess calculation based on transient analysis does not
work with the AFM based HB method. The transient analysis runs first
to get the time domain initial guess. It is then converted to
frequency domain initial guess using certain FT. The multi-tone HB
and single HB use different FT. For the AFM method, it uses FFT. But
the FFT does not relate the time domain and frequency domain information.
So the current handling of initial guess calculation does not work.
The commit fixes the initial guess for AFM HB