Add better handling of whitespace in contexts
Issue #229
"Parsing error when V() expression contains extraneous trailing spaces"

In various contexts, the expression lexer was not handling whitespace
very well.  This commit addresses that issue. It addresses it by (1)
modifying the regex of certain rules to allow for arbitrary whitespace,
and (2) then removing that whitespace from the string that is placed on
the stack.