Add OpenMP search to CMake build system
In looking at more recent Trilinos features (in particular, those that
use Kokkos), Xyce needs to be able to link to OpenMP. This adds that
capability, but does the search only if the `Kokkos_DEVICES` list from
Trilinos contains `OPENMP`. We will have to pay attention to whether
OpenMP ever gets used outside of Kokkos (it shouldn't, but that doesn't
mean it won't). If it does, then the build of Xyce will break, possibly
even at run time.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/182