Document .STEP with .FOUR
This commit updates the Users' Guide, Reference Guide and Xyce 7.4
Release Notes with information on .STEP with .FOUR.  A key point
is that the output file names, for the .FOUR output, have changed.
They are are now <netlistName>.fourX where X is the step number.
This is now consistent with the file naming convention for the
.MEASURE and .FFT output.

Issues #228
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/228