Improve -o Support for .MEASURE, .FFT, .FOUR and .STEP
This commit modifies the effect of the -o command line on the
file names produced by .STEP analyses and by .MEASURE, .FOUR
and .FFT lines.  The requested -o filename (e.g., dashoFilename)
is now used as the "base filename" for those outputs rather than
the netlistName.  So, this command line:

  Xyce -o dashoFilename netlistName

will now produce the following output files for a .TRAN analyses
that has also has a .MEASURE TRAN statement in the netlist:

  dashoFilename  (with the results requested on the .PRINT TRAN lines)
  dashoFilename.mt0

If a .STEP analyses is done with two steps then the output file names
will now be this:

  dashoFilename  (with the results requested on the .PRINT TRAN lines)
  dashoFilename.res  (with the step information in it)
  dashoFilename.mt0
  dashoFilename.mt1

Similar rules now also apply to .FOUR, .FFT and .MEASURE TRAN_CONT
lines, as well as for the other measure modes used for .AC, .DC
and .NOISE analyses.

The technical approach is to pass a const reference to the CmdParse
objects into the appropriate managers.  Those managers' output
functions can then call a common set of functions that use the
information stored in the CmdParse object.  This approach should
allow this filename change to also be implemented for the -r command
line option, if needed, in a centralized place in the code.
Finally, the new functions in N_IO_CmdParse.C do prevent the user
from accidentally overwriting the netlist file name.  (For .PRINT
lines, that was SRN Bug 2007.)

Issue #222
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/222