Cleanup, removing some string copies/processing
Issue #229
"Parsing error when V() expression contains extraneous trailing spaces"

Issue #35
"Reduce netlist setup time for modern PDKs"

In my earlier commit, which fixed issue #229, I had to add string
processing to the lexer in various places, to remove white space in
different contexts.  There already was some code for removing whitespace
in the parser, and this code was now redundant.  Furthermore, that code
in the parser required some string copies.  So, this commit removes the
redundant whitespace processing from the parser, as well as some of the
string copies.