Add check instance==null to addDeviceInstance func
The DeviceMgr::addDeviceInstance function (among other things) calls
this function in the top half of the function:

  Device &device = getDeviceByModelType(model_type);
  DeviceInstance *instance = device.addInstance( ... args ...  );

The device.addInstance function is a device master class function, and
performs the allocation of an instance class.  If this function
fails for some reason, it returns a zero.  i.e., the pointer
*instance=0.

However, the second half of DeviceMgr::addDeviceInstance makes extensive
use of the instance pointer. Prior to this commit, the function never
checked if "instance" was a valid pointer. In a circuit that I was
testing today, this caused a segmentation fault, when the code
should have exited cleanly with an error.

In this commit, I added an if-statment:

if(instance!=0)

around the second half of this function.  I considered also adding an
"else" section with an error trap, but ultimately did not.  That is
because the upstream code (which calls this function) already has an
error trap, which will cause Xyce to exit with this message:

*** Xyce Abort ***
function CktNode_Dev::instantiate:
 null Device Instance pointer

So, it doesn't (at the moment) seem necessary to also have an error
trap here.  This commit makes the segfault go away.