Small optimization to DeviceMgr::getModelType func
This function has a block of code that checks for zero-resistance
resistors.  Inside that block, it checks if the resistance parameter is
expression depedent, and if it *is* expression dependent, it then checks
if it is dependent on global parameters, special variables (like TIME)
or random operators.  Only if it isn't dependent on any of these things
will then do the rest of the operation.

The check for global params and specials was much less efficient than it
needed to be.  So, this commit fixes that minor issue.