Fix R=0 resistors to allow models.
Issue #235.
"Extend the R=0 capability to allow use of resistor .MODELs"

This came up in the context of the ADE project.  The circuit I was
trying to run today contained this:

.model Rpdiff_lvsres R( r=0.1)
RESD     IN1           OUT1      Rpdiff_lvsres   0

This works if the Rpdiff_lvsres reference is removed from the RESD line.
It also works if the model is kept in the instance, but the value of the
resistor is not zero.

So, the combination of R=0 and a model name confuses Xyce.  This is
because Xyce doesn't realize that Resistor and Resistor3 devices should
be able to refer to the same kinds of models.  And Rpdiff_lvsres is
added to the device package as a Xyce::Device::Resistor type, not
Xyce::Device::Resistor3 type.

In this commit, I have fixed the issue.  The fix is to check if the type has
been automatically changed to Resistor3, and if so, to then check if
there is a specified model.  If there is a specified model, just change
its name to a NULL string.  That way, Xyce will think it was a resistor
without a model, and there will be no lookup to fail.

It is OK to do this, because the Resistor3 device won't use the model
information anyway.