Fix time step drops outs.
In time stepping from Simulink, if Xyce when past Simulink's
requested step, then The code would fill in the missin steps
with zero.  This was occuring in 3 cases: The DC op, transient
steps and potentially at the end of the simulation.  Now all
three of these cases correctly querry Xyce for the solution
to fill in the needed data point.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/104