Improve -o Support for .PRINT Lines
This commit modifies the effect of the -o command line on the
file names produced by .PRINT lines.  The requested -o filename
(e.g., dashoFilename) is now used as the "base filename" for those
outputs.  As an example, this command line:

  Xyce -o dashoFilename netlistName

will now produce the following output files for a .TRAN analysis
that has both .PRINT TRAN and .PRINT SENS statements in the netlist:

  dashoFilename.prn
  dashoFilename.SENS.prn

Both output files will use FORMAT=STD with space as the delimiter.
Any FORMAT, DELIMITER and/or FILE parameters on the individual .PRINT
and .PRINT SENS lines in the netlist will be effectively ignored
when -o is specified on the command line.  (This limitation was
needed so that "print line concatentation" is compatible with -o.)
In addition, the -delim command line option will be ignored when
-o is used.

If the -r command line option is used to generate "raw override
output" then the dashrFilename only applies to that output.  The
-o command line option (if also given) will be used as the base
filename for the other output files.  So, this command line:

  Xyce -r dashrFilename -o dashoFilename netlistName

for a netlist that also has .PRINT SENS and a .MEASURE TRAN
statements will produce these three output files:

  dashrFilename
  dashoFilename.SENS.prn
  dashoFilename.mt0

There are two other limitations on -o output.  First, the
dashoFilename can not be a directory name.  Second, if the
dashoFilename contains a directory path (<path>/dashoFilename)
then <path> must already exist.  Xyce will not create new
directories.  (Note: This second limitation also exists for
the -r, -l, -namesfile and -rsf command line options.)

Issue #222
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/222