Partial fix .options parser scale=# w/dep params
Issue #239
"Make ".options parser scale=#" play nice with dependent parameters"

This was reported by an external user.

If a netlist used ".options parser scale" to resize MOSFET sizing
parameters (like L and W), but the MOSFET models also were using
operators like AGAUSS, then the circuit most likely would not run.
It turned out that this was b/c the scaling was being applied more
than once.

There were myriad reasons why scaling was applied more than once.

The biggest issue, that is fixed in this commit, is that I was applying the scale
factors in the "processParams" function, and this function gets called
multiple times for any dependent parameter processing.  So, for example,
if "setParam" is called on ANY of the parameters in a MOSFET model, then
that model has to re-execute the process params function.   Most likely,
this then causees L,W, etc. to be "scaled" 2 or more times, and that
breaks everything, unless the scale factor was 1.0.  To address this
problem, the scaling is now applied in a different function:
Instance::applyScale.   This function is an inherited function
from the base instance class.  The "Instance::applyScale" function is
called in the device constructor, right before processParams.  This
change (which is the main focus of this commit) fixes the test circuit
provided by the external user.

Another issue, that isn't yet fixed is how to handle length scaling when
sampling is applied directly to a sizing parameter such as L, W.  I have
started to work on that, but that aspect isn't complete in this commit.

Another, smaller issue (not fixed in this commit) was that the
AGAUSS operator is unconditionally treated like a dependent (.global)
param, even if sampling is not being applied.   That is a smaller issue,
but it should get fixed.    Until it does, "processParams" will get
called more than it should in circuits using AGAUSS but not using
sampling.  That won't break the code, but it will cause it to run more
slowly.