Fix use of "options parser scale" with sampling
Issue #239
"Make ".options parser scale=#" play nice with dependent parameters"

This is a second commit on this topic, and probably the final one for
now.

The feature "option scale" is an Hspice option, which scales area
parameters in MOSFET models.  This allows the user to specify parameters
like L,W in scaled units.

I started implementing scale a while ago, but it became apparent
recently that I made some mistakes, which in some circumstances caused
the scale factor to be applied more than once.  Generally, this was
happening whenever the "processParams" functions of a MOSFET model
were forced to be evaluated multiple times.  The generic case, which I
fixed yesterday was easy enough to fix.  In that case, I just needed to
pull the scalar code out of the processParams functions, and this
prevented that scalar code from being executed more than once.

However, that fix did NOT address the use case of modifying scaled
parameters via .STEP or sampling.  If you are updating a geometric
parameter, then the code has to be smart enough to apply the scalar
every time it is updated.  That is the use case addressed in this
commit.

To fix it, I did a few things.

1.  The "Depend" structure, which the device package uses the track
    attributes of parameters, has been expanded to include both a length
    scalar dependence boolean and also a area scalar dependence boolean.

2.  In the device entity, I modified the DeviceEntity::setParam function
    and also the DeviceEntity::updateGlobalAndDependentParameters
    function.   These functions were modified to check any updated
    parameter to see if it is subject to length or area scaling, and if so,
    apply the scaling.

3.  The BSIM4 device needed to be hacked a bit to make this work
    properly; for some reason when I applied scaling to the L,W
    parameters in the BSIM4, I was doing it via the setParam function, which
    had the effect of apply the scalar more than once in this new system.
    It wasn't necessary to do this, so I've fixed that.

The one thing that this commit does NOT do, and which was outside the
scope of this bug, was to re-bin models when L, W are changed.  Getting
that to work completely correctly will require more refactoring, and
should be the subject of another, separate issue.