Adds a check in the parseIncludeFile method for badly-formed subcircuits
This commit checks if subcircuits in hierarchical netlists are well formed.
Xyce has been testing that subcircuits in flattened netlists are well formed,
but has not been testing hierarchical netlists.  This caused an issue, an
infinite loop, in the parsing of a very large ASIC.  The check ensures that
the context at the beginning of the parseIncludeFile method is the same as
at the end of the method.  If the context is not the same, then the include
file contained a subcircuit that was not terminated properly, with '.ENDS'.
An error is now thrown indicating that issue.

If this method exits with the context in a different place than when it
entered, the behavior of parsing is undefined.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/245