Add method to enable frequency domain loading of B vector
The previous HB algorithms use only time domain loading of B
vector from sources. This does not work for the AFM based HB algorithm.
The commit adds method to set B vector loading in frequency or time
domain. The B vector from sources can be loaded either in time domain
or frequency domain, but not both. Not all types of sources are supported
for frequency domain loading. The default B vector loading is
in time domain for sources

Xyce#219