Silence a reorder warning
This commit silences a re-ordering warning from the compiler.
It was introduced about a month ago with commit 8fd6a82d. I just
noticed it this morning.  Various of Tom's commit messages (on this
topic, with the same commit title) have good discussions of this
type of warning.  I won't repeat them here.