Add optional subckt separator, '.' instead of ':'
Issue #223
"Add a new subcircuit separator, "." instead of ":" as part of
-hspice-ext"

This was suggested by an external user.

For years, Xyce has used the colon, ":" as a separator for subcircuit
names and also parameter names (a la R1:R).  This was largely inspired
by what we found in Spice3 back in 1999.

However, very few codes use a colon as a separator.  Most use a period,
"." instead.  This includes ngspice, which was derived from Spice3.  But
it also includes most commercial codes like Hspice.  As the external
user said, in response to the ":" separator:  "That is incredibly
non-standard..."

It is easy to see why most codes have adopted a period.  Using a colon
makes the use of ternary expression operators much more problematic.

Because of this issue, the ternary operator, in both the old and new
expression libraries, has to have strategically placed spaces in some
expressions that use ternary operators.   This, unfortunately, makes
it harder for the expression library to deal with white space in general.
Instead of unilaterally getting rid of whitespace in expressions prior
to parsing, the expression parser has to be smart enough to only
eliminate whitespace some of the time.  This makes the grammar more
complex a slightly more error-prone.

Anyway, for the purposes of compatibility, and also looking ahead to
simplifying the expression library grammar, I started looking at this
issue on Friday.   I added support for an alternative separator in the
N_IO_ParserMgr class, as well as the N_ULT_HpsiceBools.h heaader.

There is now a global variable, called Xyce::Util::separator.  It is of
the "char" type.  By default it is ':'.  But it can also be '.'.
Currently, the only way to set it to '.' is using the "hspice" command
line options.  Either of these will set it:

Xyce -hspice-ext all
Xyce -hspice-ext separator

This required me to go thru a LOT of the source code, and find all the
places that ':' or ":" was being used in variable names.  This turned
out to be a lot of places and it wasn't straightforward to grep for
this.  So, it took a while to find all of them.