Add #include<algorithm> to this file
My recent changes to this file included the use of std::replace.

This did not break the build on my OSX laptop, but it broke
several other builds, including my Linux workstation (RHEL7, using g++).

Adding #include<algorithm> to this file fixes the problem.

This is related to Issue #223:
"Add a new subcircuit separator, "." instead of ":" as part of -hspice-ext"