Remove one use of separator
Issue #223
"Add a new subcircuit separator, "." instead of ":" as part of
-hspice-ext"

I discovered this today as I was working on tests for the new period
separator option.

In one use case, the code was applying the subcircuit separator 2x.
This redundant use was not caused (AFAIK) by my recent changes; it seems
to have been there for a while.  But, it did not matter b/c it mostly
was only used internally, and the internal code was consistent with
itself.

The use case was for model names, for .MODEL cards found inside of
subcircuits.  If I want to print a parameter from one of these models, I
have to fully resolve the name on the .print line.  And, some of the
time, I could not find a name that worked.  It turned out, that some of
the time, the separator was used 2x, so when I would use
"subcircuit:model:param" (or "subcircuit.model.param") the code would
expect "subcircuit::model:param" (or "subcircuit..model.param").

The code was applying the separator in the CircuitContext::findModel
function, and also in the DistToolBase::instantiateDevice function.

It turns out the the test suite had a single test affected by this
issue, the OP/sand25OpAlone.cir test.