Allow the source devices to set time to support frequency domain loading
The time in a simulation is normally set by the time integration
package or by the loader. The devices can access the time information
using solver state. But they cannot change the time. To load some
types of the source devices, such as pulse, in frequency domain, a
time domain sampling is needed. This requires the device to set
the time instead of getting it from the solver state. The frequency
domain loading of source devices is needed by the frequency mapping based
multi-tone HB methods. This commit adds methods to allow the source
devices to set time to support B vector loading in frequency domain