Fix mistakes in BSIM-SOI error messages
The BSIM-SOI tries to do some very strange things with optional nodes
when "TNODEOUT" is specified.  Specifically, TNODEOUT=1 means "the
last node specified is the temperature" irrespective of how many nodes
are actually specified.

The way they have tried to do it in the BSIM-SOI 4.5 is absolutely,
positively impossible to do in Xyce, because their Verilog assumes
that it is possible to tell the difference between specified nodes in
some way other than their position on the netlist line.  The way they
do it in BSIM-SOI 4.6 is not quite as bad, but still wrong and
impossible to do in Xyce.

Thus, in the Xyce implementation, only TNODEOUT=0 is supported with
optional nodes, with the temperature node only being available when
all 7 are specified.  TNODEOUT=1 is only supported when all 7 nodes
are specified, too (making it completely pointless).

There were mistakes in the error messages that were bad enough that
any user who got those messages would have no idea what they did
wrong.

This commit makes the error message actually say why it's being thrown
in sufficient detail that the user could actually fix the problem.

Issue was never opened in an issue tracker, this resulted from private
email with Garrick Ng.