Fix frequency comparision in frequency domain loading of Pulse source
The frequency domain loading of pulse source is based on single
tone FFT. This information needs to be correctly loaded to the
frequency spectrum of the multi-tone HB simulation. Frequency comparison
is involved in the process. The frequencies are normally very large
and the numerical errors are unlikely to be smaller than the machine
precision. This can lead to the wrong results. The error tolerance
should be frequency dependant. It should also depends on machine
precision. This commit fixes the issue.

Xyce#219