Move lead current enable to late initialization
In the original design of lead currents, a "lead current request" was
made at netlist parsing time, which simply added a device name to a
vector in the device manager.  Actually enabling the lead currents in
devices was done at the time the devices were constructed in the
addDeviceInstance call of the device manager.

Doing it this way makes it difficult to handle lead current wildcards
in external output (that using the ExtOutInterface) because these
requests happen in between initializeEarly and initializeLate --- that
is, after devices have been constructed and after initial topology
calls, but before external stuff (devices, output) are actually final.

This commit removes the calls to enableLeadCurrentCalc() on devices
until right before topology actually allocates vectors and registers
LIDs --- that is, in initializeLate.

This commit does not fix issue 243, but is a necessary first step
before I can even think of fixing up the wild card issue.

Xyce#243
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/243
https://github.com/Xyce/Xyce/issues/29